requirements.txt
gradio
python-docx
deep-translator
httpcore
googletrans
lxml